一日のカロリー、タンパク質量表示・追加

+++ lib/set_page/set_controller.dart
@@ -22,8 +22,8 @@ class SetController extends ChangeNotifier {
   }
 
   Future<void> setUserDate({
-    required double totalCalorie,
-    required double totalProtein,
+    required int totalCalorie,
+    required int totalProtein,
     required double weight,
     required String userName,
   }) async {
